pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-graph-env"
version = "0.1.0"
description = "Social Graph Manipulation Detection Environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "pydantic>=2.0.0",
    "networkx>=3.0",
    "Faker>=19.0.0",
    "openai>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
package-dir = { "server" = "server" }